runtime: move inject queue to task module
